requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=mkl
brotlipy=0.7.0=py311h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2023.7.22=h56e8100_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py311h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py311h21b164f_0
cuda-cccl=12.1.109=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.1.105=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pypi_0
debugpy=1.6.7=py311hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
efficientnet-pytorch=0.7.1=pypi_0
executing=1.2.0=pyhd8ed1ab_0
filelock=3.9.0=py311haa95532_0
fonttools=4.39.4=pypi_0
freetype=2.12.1=ha860e81_0
fsspec=2023.6.0=pypi_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
huggingface-hub=0.16.4=pypi_0
idna=3.4=py311haa95532_0
imageio=2.31.1=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46319
ipykernel=6.25.0=pyh6817e22_0
ipython=8.14.0=pyh08f2357_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=py311haa95532_0
joblib=1.3.1=pypi_0
jpeg=9e=h2bbff1b_1
jupyter_client=8.3.0=pyhd8ed1ab_0
jupyter_core=5.3.1=py311h1ea47a8_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.3=pypi_0
lerc=3.0=hd77b12b_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.2.106=0
libcurand-dev=10.3.2.106=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py311h2bbff1b_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h8bd8f75_46356
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.6=py311hf62ec03_1
mkl_random=1.2.2=py311hf62ec03_1
mpmath=1.2.1=py311haa95532_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=2.8.4=py311haa95532_1
numpy=1.24.3=py311hdab7c0b_1
numpy-base=1.24.3=py311hd01c5d8_1
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1v=h2bbff1b_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.2=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.4.0=py311hd77b12b_0
pip=23.0.1=py311haa95532_0
platformdirs=2.5.2=py311haa95532_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.5=pypi_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.16.1=pyhd8ed1ab_0
pyopenssl=23.0.0=py311haa95532_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py311haa95532_0
python=3.11.3=h966fe2a_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.11=2_cp311
pytorch=2.0.1=py3.11_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pywin32=305=py311h2bbff1b_0
pyyaml=6.0=pypi_0
pyzmq=25.1.0=py311hd77b12b_0
requests=2.29.0=py311haa95532_0
safetensors=0.3.1=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=67.8.0=py311haa95532_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=py311haa95532_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.7.18=pypi_0
timm=0.9.5=pypi_0
tk=8.6.12=h2bbff1b_0
torchaudio=2.0.2=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.3.2=py311h2bbff1b_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.5.0=py311haa95532_0
tzdata=2023.3=pypi_0
ultralytics=8.0.114=pypi_0
urllib3=1.26.15=py311haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
xz=5.4.2=h8cc25b3_0
zeromq=4.3.4=h0e60522_1
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
